+++ app/src/main/java/com/sideral/ovejas/model/HeaderItem.kt
@@ -1,7 +1,7 @@
 package com.sideral.ovejas.model
 
 class HeaderItem(title: String) : Item(title){
-    override fun getItemType(): Short{
+    override fun getItemType(): Int{
         return HEADER_ITEM_TYPE;
     }
 }
