+++ app/src/main/java/com/sideral/ovejas/model/Item.kt
@@ -1,10 +1,10 @@
 package com.sideral.ovejas.model
 
 abstract class Item(title: String) {
-    var HEADER_ITEM_TYPE : Short = 0
-    var GRID_ITEM_TYPE : Short = 1
+    var HEADER_ITEM_TYPE : Int = 0
+    var GRID_ITEM_TYPE : Int = 1
     var mItemTitle = title
 
-    abstract fun getItemType(): Short
+    abstract fun getItemType(): Int
 
 }
